refactor: rename plan route components to kebab-case
Rename the eight PascalCase route-local components under
src/routes/(app)/plan/[id]/ to kebab-case filenames per the naming
convention in CLAUDE.md, and update the import paths in +page.svelte
and cash-edit-dialog.svelte. Component names stay PascalCase; file
contents are otherwise unchanged.

Closes #127
