+++ src/routes/(app)/plan/[id]/asset-edit-dialog.svelte
