+++ src/routes/(app)/plan/[id]/plan-sidebar-row.svelte
